consumer/requirements.txt
paho-mqtt
geopy
diskcache
